requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
